// projects/mf-components/src/utils/models/global.model.ts
export enum EventTypeTrigger {
  SUBMIT = 'submit',
  FETCHING = 'fetching',
  FILL_DATATABLE = 'fillDataTable',
}

export enum EventTypeElementsEnum {
  ONCLCIK = 'onClick',
  ONINIT = 'onInit',
  ONCHANGE = 'onChange',
}

export class SelectionModel {
  key: string;
  label: string;
  constructor(item?: SelectionModel) {
    this.key = item?.key ?? '';
    this.label = item?.label ?? '';
  }
}

// TODO: Check with the enum EventTypeElementsEnum
export type EventTypeElement = 'onClick' | 'onInit' | 'onChange';

export type FormFieldValue = {
  [key: string]: any;
};

export class GeneralSchemaFormBuilder {
  fields: FormSchemaField[];
  schemeColumns: FormSchemaLayoutColumn[];
  generalSetting: FormSchemaGeneralSettings;
  variables: string[]
  fetching?: FormSchemaFetching;
  constructor(item?: GeneralSchemaFormBuilder) {
      this.fields = item?.fields ?? [];
      this.schemeColumns = item?.schemeColumns ?? [];
      this.generalSetting = item?.generalSetting ?? new FormSchemaGeneralSettings();
      this.variables = item?.variables ?? [];
      this.fetching = item?.fetching ?? {};
  }
}

export class FormSchemaLayout {
  key: string;
  numberColumns?: number;
  schemaColums: Array<string>;

  constructor(item?: FormSchemaLayout) {
      this.key = item?.key ?? ''
      this.numberColumns = item?.numberColumns ?? 0
      this.schemaColums = item?.schemaColums ?? [];
  }
}

export class FormSchemaLayoutColumn {
  key: string;
  numberColumns: number;
  schemaLayout: Array<FormSchemaLayout>;
  constructor(item?: FormSchemaLayoutColumn) {
      this.key = item?.key ?? ''
      this.numberColumns = item?.numberColumns ?? 0;
      this.schemaLayout = item?.schemaLayout ?? [];
  }
}

export interface FormSchemaPropertyStyle {
  position?: ElementPosition;
  dimensions?: Dimension;
  style?: MobileformsCSSStyleProperties;
}

export class FormSchemaField {
  key: string;
  isArray?: boolean;
  type: string;
  label?: string;
  title?: string;
  value?: string;
  mask?: string;
  specialCharacters?: Array<string>;
  patterns?: IConfig['patterns'];
  required?: boolean;
  disabled?: boolean;
  placeholder?: string;
  options?: Array<SelectionModel>;
  multiline?: boolean;
  eventType?: string;

  /**
   * Table properties
   */
  header?: TableHeader<any>[];
  records?: Array<any>;

  /**
   * Dropdown Element
   */
  optionsVariable?: string;

  /**
   * Button Element
   */
  disableWithControls?: boolean;
  /**
   * For no control elements
   */
  typeReference?: string;
  /**
   * Experimental
   */
  propertiesStyle: FormSchemaPropertyStyle;
  data?: DataFormSchemaField;
  fetching?: FormSchemaFetching

  constructor(item?: FormSchemaField) {
    this.key = item?.key ?? '';
    this.isArray = item?.isArray ?? false;
    this.type = item?.type ?? '';
    this.label = item?.label ?? '';
    this.title = item?.title ?? '';
    this.value = item?.value ?? '';
    this.mask = item?.mask ?? '';
    this.specialCharacters = item?.specialCharacters ?? [];
    this.patterns = item?.patterns ?? {};
    this.required = item?.required ?? false;
    this.disabled = item?.disabled ?? false;
    this.placeholder = item?.placeholder ?? '';
    this.options = item?.options ?? [];
    this.multiline = item?.multiline ?? false;
    this.eventType = item?.eventType ?? '';
    this.propertiesStyle = item?.propertiesStyle ?? {};
    this.typeReference = item?.typeReference ?? '';
    this.header = item?.header ?? [];
    this.fetching = item?.fetching ?? {};
    this.optionsVariable = item?.optionsVariable ?? '';
  }
}

export interface FormSchemaFetching {
  typEvent?: EventTypeElement | null;
  urlFetching?: string;
  variable?: string;
  method?: MethodFetching;
  keys?: {[key: string]: boolean}
}


export interface FormSchemaFetchingWithIdentifierElement extends FormSchemaFetching {
  identifierElement: string;
}


export class DataFormSchemaField {
  fieldReference: DataFieldReference;
  constructor(item?: DataFormSchemaField) {
    this.fieldReference = item?.fieldReference ?? {} as DataFieldReference
  }
}

export interface DataFieldReference {
  keyTable: string;
  reference: DataFielReference[];
}

export interface DataFielReference {
  columnReference: string;
  fieldReference: string;
}

export class FormSchemaGeneralSettings {
  endpointUrlBase: string;
  token: string;
  typeForm?: string;

  constructor(item?: FormSchemaGeneralSettings) {
    this.endpointUrlBase = item?.endpointUrlBase ?? '';
    this.token = item?.token ?? '';
    this.typeForm = item?.typeForm ?? '';
  }
}


export type MethodFetching = 'GET' | 'POST' | 'PUT' | 'DELETE'
export interface FetchingConfig{
    baseUrl: string;
    endpoint: string;
    paramUrl?: string;
    data?: Record<string, any>
    method: MethodFetching;
    headers?: Record<string, string>
}

/**
 * Experimental
 */

export interface ElementPosition {
  x: number;
  y: number; 
}


export interface ElementsTheming {
    [type_element: string]: MobileformsCSSStyleProperties
}